# goblas: optimize Ddot using SSE2 on amd64