--- a/goblas/ddot.go
+++ b/goblas/ddot.go
@@ -1,3 +1,7 @@
+// Copyright ©2014 The gonum Authors. All rights reserved.
+// Use of this source code is governed by a BSD-style
+// license that can be found in the LICENSE file.
+
 // +build !amd64
 
 package goblas