--- a/goblas/ddot_amd64.go
+++ b/goblas/ddot_amd64.go
@@ -1,0 +1,4 @@
+package goblas
+
+func ddotUnitary(x, y []float64) (sum float64)
+func ddotInc(x, y []float64, n, incX, incY, ix, iy uintptr) (sum float64)