--- a/goblas/ddot_amd64.go
+++ b/goblas/ddot_amd64.go
@@ -1,3 +1,7 @@
+// Copyright ©2014 The gonum Authors. All rights reserved.
+// Use of this source code is governed by a BSD-style
+// license that can be found in the LICENSE file.
+
 package goblas
 
 func ddotUnitary(x, y []float64) (sum float64)