--- a/goblas/ddot_amd64.s
+++ b/goblas/ddot_amd64.s
@@ -1,3 +1,7 @@
+// Copyright ©2014 The gonum Authors. All rights reserved.
+// Use of this source code is governed by a BSD-style
+// license that can be found in the LICENSE file.
+//
 // Some of the loop unrolling code is copied from:
 // http://golang.org/src/math/big/arith_amd64.s
 // which is distributed under these terms: