--- a/goblas/level1double.go
+++ b/goblas/level1double.go
@@ -1,3 +1,7 @@
+// Copyright ©2014 The gonum Authors. All rights reserved.
+// Use of this source code is governed by a BSD-style
+// license that can be found in the LICENSE file.
+
 // Uses the netlib standard. Other implementations may differ. Difference
 // is that the code panics for n < 0 and incx == 0 rather than returning zero.
 // (Documentation says incx must not be zero)