@kortschak 2015-01-01
I get this:

```
$ benchcmp old.txt new.txt 
benchmark                          old ns/op     new ns/op     delta
BenchmarkDdotSmallBothUnitary      23.1          21.6          -6.49%
BenchmarkDdotSmallIncUni           36.3          29.5          -18.73%
BenchmarkDdotSmallUniInc           44.9          28.0          -37.64%
BenchmarkDdotSmallBothInc          37.8          27.6          -26.98%
BenchmarkDdotMediumBothUnitary     1627          627           -61.46%
BenchmarkDdotMediumIncUni          2344          1647          -29.74%
BenchmarkDdotMediumUniInc          2478          1618          -34.71%
BenchmarkDdotMediumBothInc         2504          1945          -22.32%
BenchmarkDdotLargeBothUnitary      186890        84719         -54.67%
BenchmarkDdotLargeIncUni           713870        634351        -11.14%
BenchmarkDdotLargeUniInc           433818        317105        -26.90%
BenchmarkDdotLargeBothInc          935882        841664        -10.07%
BenchmarkDdotHugeBothUnitary       27035479      22361410      -17.29%
BenchmarkDdotHugeIncUni            72098861      65191729      -9.58%
BenchmarkDdotHugeUniInc            48003429      50984946      +6.21%
BenchmarkDdotHugeBothInc           86445132      91902762      +6.31%
```

I can't see why the release test didn't build on travis - textflag.h should be present.

Would you add license headers?