@kortschak 2015-01-01
... and with go version devel +a877e81.

```
$ benchcmp old.txt new.txt 
benchmark                          old ns/op     new ns/op     delta
BenchmarkDdotSmallBothUnitary      22.2          20.6          -7.21%
BenchmarkDdotSmallIncUni           34.9          27.4          -21.49%
BenchmarkDdotSmallUniInc           35.3          27.5          -22.10%
BenchmarkDdotSmallBothInc          35.4          27.3          -22.88%
BenchmarkDdotMediumBothUnitary     1600          620           -61.25%
BenchmarkDdotMediumIncUni          2320          1630          -29.74%
BenchmarkDdotMediumUniInc          2400          1603          -33.21%
BenchmarkDdotMediumBothInc         2296          1946          -15.24%
BenchmarkDdotLargeBothUnitary      162103        81675         -49.62%
BenchmarkDdotLargeIncUni           626682        583883        -6.83%
BenchmarkDdotLargeUniInc           335803        298583        -11.08%
BenchmarkDdotLargeBothInc          848481        825154        -2.75%
BenchmarkDdotHugeBothUnitary       23342114      21972418      -5.87%
BenchmarkDdotHugeIncUni            66768363      63730296      -4.55%
BenchmarkDdotHugeUniInc            45670037      44210559      -3.20%
BenchmarkDdotHugeBothInc           85100256      83724079      -1.62%
```