@jonlawlor 2015-01-01
The problem is that the travis-ci release uses Go 1.3.3, while the tip uses Go 1.4.  If you look at the build details, you'll see that tip passed, but release failed.  I switched my computer to 1.3.3 and tried it out and I get the same error as the travis build.