@btracey 2015-01-01
Thanks for putting this together. A couple of high-level comments:
1) I'm not sure what we want to do about codereview/ maintenance. I can sort of tell what's going on, but that's a lot because of the nice comments in the routine. I like fast code as much as anyone, but we need to have people that can maintain the code (for example, to support changes in the runtime).

2) We'll need to increase the test suite, specifically to include cases with wider strides than the matrix and cases where the slices are shadowing one another by various amounts. With assembly coded by ourselves, we can't rely on the fact that the go team has done the right thing

3) Assuming it's maintainable and correct, this could be quite promising. There are a small number of loop kernels (ddot, axpy, and scale at least, but probably not too much more), and it's probably not that hard to recode the inner loops of the higher-level functions to use these. I don't know how much the lack of inline-assembly will hurt, but the go tool may improve at some point.

4) Is it possible to code these in such a way that it is extensible to more than just MULPD? The Intel Knights landing chips, for example, are moving in the direction of larger and larger vectorized units. It would be great if had code that could take advantage of that somehow. What I have in mind would be something like 

````
type SimdSize struct {
     length int
     mulInstruction string
     // etc.
}
````

We then have a code generator that takes in an array of SimdSizes and generates the assembly. To generate the amd64 code, you would give it only SimdSize{2, "MULPD"} while for knights landing  x87 you would give it the instructions for 128, 256, and 512 bit vector sizes. The generator tool would provide the logic for calling those in succession (as well as testing for shadowing, etc.). It seems that the future is headed toward manycore, and the Go language has the constructs to take advantage. It would be nice if we could also take advantage of the additional vectorization. It may be that the assemblies are so different that such an endeavor is too much effort. Just a thought.