@btracey 2015-01-01
Here are my benchmarks:

On 1.4, old vs new
````
benchmark                          old ns/op     new ns/op     delta       
BenchmarkDdotSmallBothUnitary      13.2          12.4          -6.06%      
BenchmarkDdotSmallIncUni           19.3          17.1          -11.40%     
BenchmarkDdotSmallUniInc           18.7          17.2          -8.02%      
BenchmarkDdotSmallBothInc          19.1          16.9          -11.52%     
BenchmarkDdotMediumBothUnitary     850           417           -50.94%     
BenchmarkDdotMediumIncUni          1128          987           -12.50%     
BenchmarkDdotMediumUniInc          1124          997           -11.30%     
BenchmarkDdotMediumBothInc         1171          1020          -12.89%     
BenchmarkDdotLargeBothUnitary      84471         45765         -45.82%     
BenchmarkDdotLargeIncUni           171155        143554        -16.13%     
BenchmarkDdotLargeUniInc           119249        103142        -13.51%     
BenchmarkDdotLargeBothInc          249504        242060        -2.98%      
BenchmarkDdotHugeBothUnitary       10760149      9142349       -15.04%     
BenchmarkDdotHugeIncUni            27278566      26659861      -2.27%      
BenchmarkDdotHugeUniInc            19205312      17377502      -9.52%      
BenchmarkDdotHugeBothInc           35331774      35384487      +0.15%
```` 

New. txt vs cblas using OSX Accelerate framework. 
````
benchmark                          old ns/op     new ns/op     delta         
BenchmarkDdotSmallBothUnitary      12.4          196           +1480.65%     
BenchmarkDdotSmallIncUni           17.1          201           +1075.44%     
BenchmarkDdotSmallUniInc           17.2          200           +1062.79%     
BenchmarkDdotSmallBothInc          16.9          200           +1083.43%     
BenchmarkDdotMediumBothUnitary     417           340           -18.47%       
BenchmarkDdotMediumIncUni          987           866           -12.26%       
BenchmarkDdotMediumUniInc          997           774           -22.37%       
BenchmarkDdotMediumBothInc         1020          1013          -0.69%        
BenchmarkDdotLargeBothUnitary      45765         49874         +8.98%        
BenchmarkDdotLargeIncUni           143554        147430        +2.70%        
BenchmarkDdotLargeUniInc           103142        87223         -15.43%       
BenchmarkDdotLargeBothInc          242060        232011        -4.15%        
BenchmarkDdotHugeBothUnitary       9142349       9228276       +0.94%        
BenchmarkDdotHugeIncUni            26659861      28009570      +5.06%        
BenchmarkDdotHugeUniInc            17377502      17337081      -0.23%        
BenchmarkDdotHugeBothInc           35384487      34871204      -1.45%  
````
Accelerate is still doing something better (especially considering the cgo overhead), but this eliminates much of the gap.