@jonlawlor 2015-01-01
Atom 510, x86_64:
```
benchmark                          old ns/op     new ns/op     delta
BenchmarkDdotSmallBothUnitary      139           113           -18.71%
BenchmarkDdotSmallIncUni           180           148           -17.78%
BenchmarkDdotSmallUniInc           201           159           -20.90%
BenchmarkDdotSmallBothInc          189           158           -16.40%
BenchmarkDdotMediumBothUnitary     8534          4918          -42.37%
BenchmarkDdotMediumIncUni          15988         9994          -37.49%
BenchmarkDdotMediumUniInc          16055         9446          -41.16%
BenchmarkDdotMediumBothInc         18266         12412         -32.05%
BenchmarkDdotLargeBothUnitary      1020829       646607        -36.66%
BenchmarkDdotLargeIncUni           1793379       1307234       -27.11%
BenchmarkDdotLargeUniInc           1746168       1135265       -34.99%
BenchmarkDdotLargeBothInc          2114698       1673166       -20.88%
BenchmarkDdotHugeBothUnitary       101451737     64170337      -36.75%
BenchmarkDdotHugeIncUni            178849842     129829209     -27.41%
BenchmarkDdotHugeUniInc            173484821     111482658     -35.74%
BenchmarkDdotHugeBothInc           209515105     168920310     -19.38%
```